run_serial: make sure to not overwrite the directories that have been already created

+++ PYSPEC/run_serial.py
@@ -20,11 +20,12 @@ import subprocess
 import sys
 
 # ------------------- INPUT -----------------------------
-req_phase = 'Pdiff'
+req_phase = 'P'
 req_solver = 'yspec'
 req_processes = '40'
+update_all = False
 add_event_info = '/import/neptun-radler/hosseini-downloads/KASRA/SCRIPTS/gitHUB/myrepo_gitHUB/FFM/INITIALIZATION/results'
-add_proc_ev = '/import/neptun-helles/hosseini/FFM_RESULTS/Pdiff_measure_1_sec_LAMBDA_1-5_90_180'
+add_proc_ev = '/import/neptun-helles/hosseini/FFM_RESULTS/P_measure_1_sec_LAMBDA_1-5_32_100'
 add_runff = '/home/hosseini/FFINVERSION/AMPLITUDES/Programs/ffproc/FFsetup'
 # -------------------------------------------------------
 
@@ -32,14 +33,18 @@ exit_flag = raw_input('Did you set run_serial=1 in RunFFProcessing? (y/n)\n')
 if exit_flag.upper() != 'Y':
     sys.exit('Please do it first! it is at the top of the file in input section!')
 
-fio_event_info = open(os.path.join(add_event_info, 'selected_events.txt'))
+fio_event_info = open(os.path.join(add_event_info, 'selected_events_ALL.txt'))
 event_info = fio_event_info.readlines()
 
 for i in range(len(event_info)):
     event_name = event_info[i].split(',')[0]
     event_process = os.path.join(add_proc_ev, event_name)
     if os.path.isdir(event_process):
-        shutil.rmtree(event_process)
+        if update_all:
+            shutil.rmtree(event_process)
+        else:
+            print "%s has been measured before..." % event_info[i]
+            continue
     os.mkdir(event_process)
     os.chdir(event_process)
     subprocess.check_call([add_runff, req_phase, req_solver, req_processes])
